+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- range slider: make track color gray below first thumb and above last thumb (configurable through `outerTrackColor`).
+
 ## [1.8.2] - 2018-11-06
 
 ### Fixed